+++ requirements.txt
@@ -1,1 +1,3 @@
 requests==2.24.0
+pytest~=6.1.1
+setuptools~=50.3.2
