+++ tests/conftest.py
@@ -1,1 +1,1 @@
-pytest_plugins = ['pytester']
+pytest_plugins = ['pytester', 'europy']
